# backend/PandaScooter.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PandaScooter",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/map": {
      "get": {
        "summary": "查询车辆，禁停区，停放点",
        "deprecated": false,
        "description": "查询附近车辆",
        "tags": [],
        "parameters": [
          {
            "name": "longitude",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "latitude",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "scale",
            "in": "query",
            "description": "地图缩放级别 3-20",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "scooters": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "description": "scooter主键ID",
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 2147483647
                              },
                              "code": {
                                "description": "车辆编码",
                                "type": "string",
                                "maxLength": 32
                              },
                              "ride_status": {
                                "description": "状态：0-空闲，1-使用中",
                                "default": "0",
                                "type": "integer",
                                "minimum": -2147483648,
                                "maximum": 2147483647,
                                "nullable": true
                              },
                              "fault_status": {
                                "type": "integer",
                                "description": "0-正常 1-故障"
                              },
                              "battery": {
                                "description": "电量",
                                "default": "100",
                                "type": "integer",
                                "minimum": -2147483648,
                                "maximum": 2147483647,
                                "nullable": true
                              },
                              "latitude": {
                                "description": "纬度",
                                "type": "number",
                                "nullable": true
                              },
                              "longitude": {
                                "description": "经度",
                                "type": "number",
                                "nullable": true
                              }
                            },
                            "required": [
                              "id",
                              "code",
                              "fault_status"
                            ]
                          }
                        },
                        "noParkingAreas": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "description": "禁停区主键ID",
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 2147483647
                              },
                              "polygon": {
                                "description": "禁停多边形范围坐标",
                                "type": "string",
                                "maxLength": 65535,
                                "nullable": true
                              },
                              "status": {
                                "description": "状态：1-启用，0-禁用",
                                "default": "1",
                                "type": "integer",
                                "minimum": -2147483648,
                                "maximum": 2147483647,
                                "nullable": true
                              },
                              "center": {
                                "type": "string",
                                "description": "禁停多边形中心点坐标"
                              }
                            },
                            "required": [
                              "id",
                              "center"
                            ]
                          }
                        },
                        "parkingPoints": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "latitude": {
                                "type": "number"
                              },
                              "longitude": {
                                "type": "number"
                              }
                            },
                            "required": [
                              "latitude",
                              "longitude"
                            ]
                          }
                        }
                      },
                      "required": [
                        "scooters",
                        "noParkingAreas",
                        "parkingPoints"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/scooter/unlock": {
      "post": {
        "summary": "解锁小车开始订单",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "orderId": {
                          "type": "integer"
                        },
                        "scooterId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "orderId",
                        "scooterId"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/scooter": {
      "get": {
        "summary": "扫码获取小车信息",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "description": "主键ID",
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 2147483647
                        },
                        "code": {
                          "description": "车辆编码",
                          "type": "string",
                          "maxLength": 32
                        },
                        "ride_status": {
                          "description": "状态：0-空闲，1-使用中",
                          "default": "0",
                          "type": "integer",
                          "minimum": -2147483648,
                          "maximum": 2147483647,
                          "nullable": true
                        },
                        "fault_status": {
                          "type": "integer",
                          "description": "0-正常 1-故障"
                        },
                        "battery": {
                          "description": "电量",
                          "default": "100",
                          "type": "integer",
                          "minimum": -2147483648,
                          "maximum": 2147483647,
                          "nullable": true
                        },
                        "latitude": {
                          "description": "纬度",
                          "type": "number",
                          "nullable": true
                        },
                        "longitude": {
                          "description": "经度",
                          "type": "number",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "code",
                        "fault_status"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/scooter/lock": {
      "post": {
        "summary": "锁上小车结束订单",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderId": {
                    "description": "订单ID",
                    "type": "integer",
                    "minimum": -2147483648,
                    "maximum": 2147483647
                  },
                  "startTime": {
                    "description": "开始时间",
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "endTime": {
                    "type": "string",
                    "description": "结束时间",
                    "format": "date-time",
                    "nullable": true
                  },
                  "amount": {
                    "description": "金额",
                    "default": "0.00",
                    "type": "number",
                    "nullable": true
                  },
                  "totalKilometer": {
                    "description": "总里程",
                    "default": "0.00",
                    "type": "number",
                    "nullable": true
                  },
                  "code": {
                    "description": "车辆主键ID",
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 2147483647
                  },
                  "battery": {
                    "description": "电量",
                    "default": "100",
                    "type": "integer",
                    "minimum": -2147483648,
                    "maximum": 2147483647,
                    "nullable": true
                  },
                  "latitude": {
                    "description": "纬度",
                    "type": "number",
                    "nullable": true
                  },
                  "longitude": {
                    "description": "经度",
                    "type": "number",
                    "nullable": true
                  }
                },
                "required": [
                  "orderId",
                  "code"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "description": "主键ID",
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 2147483647
                        },
                        "startTime": {
                          "description": "开始时间",
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "payStatus": {
                          "description": "支付状态：0-未支付，1-已支付",
                          "default": "0",
                          "type": "integer",
                          "minimum": -2147483648,
                          "maximum": 2147483647,
                          "nullable": true
                        },
                        "amount": {
                          "description": "金额",
                          "default": "0.00",
                          "type": "number",
                          "nullable": true
                        },
                        "totalTime": {
                          "type": "string",
                          "description": "总时长（00：10：10）十分十秒"
                        },
                        "totalKilometer": {
                          "description": "总里程",
                          "default": "0.00",
                          "type": "number",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "totalTime"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}